requirements.txt
flask
openai-whisper
pymupdf
werkzeug